+++ app/components/Views/BrowserTab/Tabs/TabCountIcon/index.js
