+++ app/components/Views/BrowserTab/Tabs/TabCountIcon/index.test.js
