+++ app/components/Views/BrowserTab/Tabs/TabModalAnimation/index.js
