+++ app/components/Views/BrowserTab/Tabs/TabModalAnimation/index.test.js
