+++ app/components/Views/BrowserTab/Tabs/TabThumbnail/index.js
