+++ app/components/Views/BrowserTab/Tabs/TabThumbnail/index.test.js
