+++ app/components/Views/BrowserTab/Tabs/index.js
