+++ app/components/Views/BrowserTab/Tabs/index.test.js
